Use heuristic to determine base package
When instrumenting a package a/b/c, we'll determine which other packages
are "relevant" to a/b/c, and therefor should also be instrumented with
a/b/c. We'll use a basepkg for that. All imports which are decendants
of the basepkg, will be instrumented with a/b/c. So if a/b is the
basepkg of a/b/c, if we'll have import to a/b/d, it'll also be
instrumented.

We'll guess the basepkg using the following heuristic:

  1. If the import path of our package is a vcs - the vcs root is the
     basepkg.
  2. Otherwise, the uppermost ancestor which is a valid go pacakge is
     the basepkg. For example, if a/b/c/d is the package we're
     instrumenting, and a/b/c is a go package, but a/b is not, then
     a/b will be the basepkg.